requirements/production.txt
# Pro-tip: Try not to put anything here. There should be no dependency in
#   production that isn't in development.
-r common.txt

# Static Files and Media Storage
# -------------------------------------
boto==2.42.0
django-storages-redux==1.3.2

# Caching
# -------------------------------------
django-redis==4.4.3
hiredis==0.2.0


# Logging
# -------------------------------------
newrelic

